refactor(metafetch): split service.go into 8 domain files
Splits the 3932-line service.go into:
- service_wiring.go (constructor, DI)
- service_fetch.go (fetch entry points)
- service_search.go (search context, source chain)
- service_apply.go (apply, persist, candidate)
- service_scoring.go (scoring, ranking)
- service_normalize.go (normalisation, parsing)
- service_writeback.go (writeback, tags, pipeline)
- service_files.go (file I/O, iTunes path)

No logic changes. Structure audit STRUCT-4.
